Renamed project module to blueprint
