+++ src/modules/blueprint/components/BlueprintListItem.vue
