+++ src/modules/blueprint/components/ExampleListItem.vue
