+++ src/modules/record/pages/list.vue
@@ -4,12 +4,12 @@
     <!-- Header -->
     <div class="col-lg-12">
       <EditorHeader :help="'Define the attributes that can be assigned to a single ' + schema.label">
-        <a class='btn btn-outline-warning float-right mr-2' :href="'#/blueprints/' + project_id >
+        <a class='btn btn-outline-warning float-right mr-2' :href="'#/blueprints/' + blueprint_id >
           <i class="fas fa-fw fa-pencil-alt mr-2"></i>
           Attributes
         </a>
 
-        <a class='btn btn-primary float-right' :href="'#/blueprints/' + project_id + '/seeds/' + schema_id + '/new'" >
+        <a class='btn btn-primary float-right' :href="'#/blueprints/' + blueprint_id + '/seeds/' + schema_id + '/new'" >
           <i class="fa fa-fw fa-plus mr-2"></i>
           New {{ schema.label }}
         </a>
@@ -18,7 +18,7 @@
 
     <!-- Record List -->
     <div class="col-lg-12">
-      <RecordList :schema="schema" :projectId="project_id" />
+      <RecordList :schema="schema" :projectId="blueprint_id" />
     </div>
 
   </div>
@@ -31,7 +31,7 @@ import { mapGetters, mapActions } from 'vuex'
 import RecordList from '@/components/record/RecordList'
 
 export default {
-  props: ['project_id', 'schema_id'],
+  props: ['blueprint_id', 'schema_id'],
   metaInfo: {
     title: 'Seed - List'
   },
