+++ src/modules/record/pages/new.vue
@@ -21,7 +21,7 @@ import { mapGetters, mapActions } from 'vuex'
 import RecordForm from '@/components/record/RecordForm'
 
 export default {
-  props: ['project_id', 'schema_id'],
+  props: ['blueprint_id', 'schema_id'],
   metaInfo: {
     title: 'Seed - New'
   },
