+++ src/store/index.js
@@ -1,7 +1,7 @@
 import createPersistedState from 'vuex-persistedstate'
 import Vuex from 'vuex'
 
-import project from '@/modules/project/store'
+import project from '@/modules/blueprint/store'
 import schema from '@/modules/schema/store'
 import attribute from '@/modules/attribute/store'
 import relation from '@/modules/relation/store'
